+++ dev/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @khanacademy/perseus-dev-ui
 
+## 1.5.2
+
+### Patch Changes
+
+-   Updated dependencies [[`b4143615b`](https://github.com/Khan/perseus/commit/b4143615b294da439339836d679052c8e1d479a5)]:
+    -   @khanacademy/math-input@18.0.1
+
 ## 1.5.1
 
 ### Patch Changes